+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [ita-challenges-frontend-3.2.1-RELEASE] - 2025-07-23
+
+### Added
+- Implemented bookmarks to the bookmarks view (Taiga [#597], PR [#674]).
+
 ### [ita-challenges-frontend-3.2.0-RELEASE] - 2025-07-23
 
 ### Added